+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/MetricUnionCompositeTable.java
@@ -5,10 +5,6 @@ package com.yahoo.bard.webservice.table;
 import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.table.availability.MetricUnionAvailability;
-import com.yahoo.bard.webservice.util.IntervalUtils;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 import java.util.Set;
@@ -55,71 +51,32 @@ import javax.validation.constraints.NotNull;
  * and this joined table is backed by the <tt>MetricUnionAvailability</tt>
  *
  */
-public class MetricUnionCompositeTable extends BasePhysicalTable {
-    private static final Logger LOG = LoggerFactory.getLogger(MetricUnionCompositeTable.class);
+public class MetricUnionCompositeTable extends BaseCompositePhysicalTable {
 
     /**
      * Constructor.
      *
      * @param name  Name that represents set of fact table names joined together
+     * @param timeGrain  The time grain of the table. The time grain has to satisfy all grains of the tables
      * @param columns  The columns for this table
-     * @param physicalTables  A set of <tt>PhysicalTable</tt>s whose same metric schema is to be joined together. The
-     * tables will be used to construct MetricUnionAvailability, as well as to compute common/coarsest time grain among
-     * them. The <tt>PhysicalTable</tt>s needs to have mutually satisfying time grains in order to calculate the
-     * common/coarsest time grain.
+     * @param physicalTables  A set of PhysicalTables that are put together under this table. The
+     * tables shall have zoned time grains that all satisfy the provided timeGrain
      * @param logicalToPhysicalColumnNames  Mappings from logical to physical names
      */
     public MetricUnionCompositeTable(
             @NotNull TableName name,
+            @NotNull ZonedTimeGrain timeGrain,
             @NotNull Set<Column> columns,
             @NotNull Set<PhysicalTable> physicalTables,
             @NotNull Map<String, String> logicalToPhysicalColumnNames
     ) {
         super(
                 name,
-                IntervalUtils.getCoarsestTimeGrain(physicalTables).orElseThrow(() -> {
-                    String message = String.format(
-                            "At least 1 physical table needs to be provided in order to calculate " +
-                                    "coarsest time grain for %s",
-                            name.asName()
-                    );
-                    LOG.error(message);
-                    return new IllegalArgumentException(message);
-                }),
+                timeGrain,
                 columns,
+                physicalTables,
                 logicalToPhysicalColumnNames,
                 new MetricUnionAvailability(physicalTables, columns)
         );
-        verifyGrainSatisfiesAllTables(getSchema().getTimeGrain(), physicalTables, name);
-    }
-
-    /**
-     * Verifies that the coarsest <tt>ZonedTimeGrain</tt> satisfies all tables.
-     *
-     * @param coarsestTimeGrain  The coarsest <tt>ZonedTimeGrain</tt> to be verified
-     * @param physicalTables  A set of <tt>PhysicalTable</tt>s whose <tt>ZonedTimeGrain</tt>s are checked to make sure
-     * they all satisfies with the given coarsest <tt>ZonedTimeGrain</tt>
-     * @param name  Name of the current <tt>MetricUnionCompositeTable</tt> that represents set of fact table names
-     * joined together
-     *
-     * @throws IllegalArgumentException when there is no mutually satisfying grain among the table's time grains
-     */
-    private static void verifyGrainSatisfiesAllTables(
-            ZonedTimeGrain coarsestTimeGrain,
-            Set<PhysicalTable> physicalTables,
-            TableName name
-    ) throws IllegalArgumentException {
-        if (!physicalTables.stream()
-                .map(PhysicalTable::getSchema)
-                .map(PhysicalTableSchema::getTimeGrain)
-                .allMatch(grain -> grain.satisfiedBy(coarsestTimeGrain))) {
-            String message = String.format("There is no mutually satisfying grain among: %s for current table " +
-                    "%s",
-                    physicalTables,
-                    name.asName()
-            );
-            LOG.error(message);
-            throw new IllegalArgumentException(message);
-        }
     }
 }